+++ cmd/selfupdate/execute/flag.go
@@ -1,0 +1,14 @@
+package execute
+
+import (
+	"github.com/spf13/cobra"
+)
+
+type flag struct {
+}
+
+func (f *flag) Init(cmd *cobra.Command) {}
+
+func (f *flag) Validate() error {
+	return nil
+}